chore: remove obsolete stubs after helper extraction
